# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: CUPHY EduTech Platform API for Calcutta University Physics Honours Students
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request OTP for phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestOtpBody"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP and login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpBody"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginWithPassword
      tags: [auth]
      summary: Login with phone and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # USERS
...
  /admin/app-settings:
    get:
      operationId: getAppSettings
      tags: [admin]
      summary: Get app settings/theme
      responses:
        "200":
          description: App settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
    put:
      operationId: updateAppSettings
      tags: [admin]
      summary: Update app settings/theme (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppSettings"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppSettings"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # AUTH SCHEMAS
    RequestOtpBody:
      type: object
      properties:
        phone:
          type: string
      required: [phone]
    VerifyOtpBody:
      type: object
      properties:
        phone:
          type: string
        otp:
          type: string
      required: [phone, otp]
    LoginBody:
      type: object
      properties:
        phone:
          type: string
        password:
          type: string
      required: [phone, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]

    # USER SCHEMAS
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, admin]
        semester:
          type: integer
        profilePicture:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, phone, role, isActive, createdAt]
...
    AppSettings:
      type: object
      properties:
        primaryColor:
          type: string
        secondaryColor:
          type: string
        logoUrl:
          type: string
          nullable: true
        appName:
          type: string
        tagline:
          type: string
          nullable: true
        darkMode:
          type: boolean
      required: [primaryColor, secondaryColor, appName, darkMode]
...