# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CUPHY EduTech Platform API for Calcutta University Physics Honours Students
 * OpenAPI spec version: 1.0.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  phone: string;
  email?: string;
  role: UserRole;
  semester?: number;
  profilePicture?: string | null;
  isActive: boolean;
  createdAt: string;
}